pyproject.toml
[project]
name = "verito"
version = "0.1.0"
description = "Forged image detection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub>=1.1.4",
    "kagglehub>=0.3.13",
    "matplotlib>=3.10.7",
    "numpy>=1.26.4",
    "pillow>=12.0.0",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.16.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "ruff>=0.14.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
